refactor: static 위치 변경
