+++ src/main/java/com/programmers/voucher/domain/VoucherFactory.java
@@ -1,0 +1,6 @@
+package com.programmers.voucher.domain;
+
+public class VoucherFactory {
+
+
+}
