+++ src/test/java/com/programmers/voucher/domain/VoucherPolicyTest.java
@@ -2,7 +2,6 @@ package com.programmers.voucher.domain;
 
 
 import com.programmers.console.exception.VoucherCommandException;
-import com.programmers.console.util.VoucherPolicy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
